(PC-33121)[PRO] fix: Create component with own state for each managed venue in pricing point dialog.
