+++ pro/src/pages/Reimbursements/BankInformations/ManagedVenueItem/ManagedVenueItem.module.scss
@@ -1,0 +1,16 @@
+@use "styles/mixins/_rem.scss" as rem;
+
+.dialog-checkbox-container {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: rem.torem(24px);
+
+  &:last-child {
+    margin-bottom: 0;
+  }
+}
+
+.dialog-checkbox-button {
+  margin: 0;
+}
